src/simapi_web/requirements.txt
amqp==2.5.2
aniso8601==7.0.0
anyjson==0.3.3
asgiref==3.2.3
billiard==3.6.3.0
celery==4.4.1
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.0.3
cycler==0.10.0
Django==3.0.3
djangorestframework==3.11.0
django-filter==2.2.0
graphene==2.1.8
graphene-django==2.9.0
graphql-core==2.2
graphql-relay==2.0.1
idna==2.8
importlib-metadata==1.3.0
itypes==1.1.0
Jinja2==2.11.1
kiwisolver==1.1.0
kombu==4.6.8
lxml==4.4.2
MarkupSafe==1.1.1
more-itertools==8.0.2
openapi-codec==1.3.2
polling2==0.4.5
promise==2.3
pyparsing==2.4.6
psycopg2-binary==2.8.4
python-dateutil==2.8.1
pytz==2019.3
redis==3.4.1
requests==2.22.0
Rx==1.6.1
simplejson==3.17.0
six==1.13.0
sqlparse==0.3.0
tornado==6.0.3
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
zipp==0.6.0
django-extensions
graphviz
sqlalchemy